requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
backoff==2.2.1
bitsandbytes==0.45.4
blake3==1.0.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.2
cupy-cuda12x==13.4.1
datasets==3.5.0
decord==0.6.0
deepspeed==0.16.5
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
flash-attn==2.6.3
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
gymnasium==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
latex2sympy2==1.5.4
latex2sympy2_extended==1.10.1
Levenshtein==0.27.1
liger_kernel==0.5.6
lightning-utilities==0.14.3
llguidance==0.7.11
llvmlite==0.43.0
lm-format-enforcer==0.10.11
loguru==0.7.3
loralib==0.1.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.7.0
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.3.2
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.70.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
-e git+https://github.com/OpenRLHF/OpenRLHF.git@db77a3019bab862747e9591001d0b3744681e079#egg=openrlhf
optimum==1.24.0
optree==0.14.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
Pebble==5.1.1
peft==0.15.1
pillow==11.1.0
platformdirs==4.3.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
pynvml==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.10
RapidFuzz==3.13.0
ray==2.44.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
ring-flash-attn==0.1.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.25.1
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tiktoken==0.9.0
timeout-decorator==0.5.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.0
transformers-stream-generator==0.0.5
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
vllm==0.8.2
wandb==0.19.9
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
word2number==1.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
